server/requirements.txt
Django==4.2.27
PyMySQL==1.1.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
Pillow==10.4.0
django-environ==0.11.2
bcrypt==5.0.0
captcha==0.7.1
python-magic==0.4.27
cryptography==42.0.0